Restructure Typescript project Configure TS compile options Update dependencies Update Blazor to 0.5.1 Move JS initialization from global to extension specific namespace
