+++ .gitignore
@@ -24,6 +24,7 @@ bld/
 [Bb]in/
 [Oo]bj/
 [Ll]og/
+dist/
 
 # Visual Studio 2015 cache/options directory
 .vs/
