+++ src/Blazor.Extensions.Logging.JS/tsconfig.json
@@ -2,11 +2,14 @@
   "compilerOptions": {
     "noImplicitAny": false,
     "noEmitOnError": true,
-    "removeComments": false,
+    "removeComments": true,
     "sourceMap": true,
-    "target": "es5",
-    "lib": ["es2015", "dom"],
-    "strict": true
+    "target": "es6",
+    "module": "commonjs",
+    "lib": ["es2016", "dom"],
+    "strict": true,
+    "alwaysStrict": true,
+    "preserveConstEnums": true
   },
   "exclude": [
     "node_modules"
