+++ src/Blazor.Extensions.Logging.JS/tslint.json
@@ -1,0 +1,15 @@
+{
+  "rules": {
+    "no-string-throw": true,
+    "no-unused-expression": true,
+    "no-duplicate-variable": true,
+    "curly": true,
+    "class-name": true,
+    "semicolon": [
+      true,
+      "always"
+    ],
+    "triple-equals": true
+  },
+  "defaultSeverity": "warning"
+}
